src/networkpolicy_manager/requirements.txt
kubernetes==17.14.0a1
requests
retry
python-slugify
PyYAML
mock